put save beatmap function inside the osudownload file
